+++ src/prog2/model/CarreteraAsfalt.java
@@ -3,9 +3,9 @@ package prog2.model;
 public class CarreteraAsfalt extends AccesAsfalt{
 
     //Atribut
-    private long pes;
+    private float pes;
 
-    public CarreteraAsfalt(String nom, boolean estat, long areaAsfalt, long pes){
+    public CarreteraAsfalt(String nom, boolean estat, float areaAsfalt, float pes){
         super(nom,estat,areaAsfalt);
         this.pes = pes;
         this.accessibilitat = true;
@@ -16,10 +16,10 @@ public class CarreteraAsfalt extends AccesAsfalt{
         return accessibilitat;
     }
 
-    public void setPes(long pes){
+    public void setPes(float pes){
         this.pes = pes;
     }
 
-    public long getPes(){return pes;}
+    public float getPes(){return pes;}
 
 }
